[#58] Refactor Test setup
